Creating new component for expense container
