+++ src/Components/ExpenseTracker/Expense/ExpenseItem.js
@@ -1,19 +1,18 @@
-import PropTypes from 'prop-types';
-import ExpenseDate from './ExpenseDate';
-import Card from '../../UI/Card';
-import './ExpenseItem.css';
+import PropTypes from "prop-types";
+import ExpenseDate from "./ExpenseDate";
+import Card from "../../UI/Card";
+import "./ExpenseItem.css";
 
 const ExpenseItem = ({ date, title, amount }) => (
-  <Card className="expense-item">
-    <ExpenseDate date={date} />
-    <div className="expense-item__description">
-      <h2>{title}</h2>
-      <div className="expense-item__amount">
-        $
-        {amount}
+  <li>
+    <Card className="expense-item">
+      <ExpenseDate date={date} />
+      <div className="expense-item__description">
+        <h2>{title}</h2>
+        <div className="expense-item__amount">${amount}</div>
       </div>
-    </div>
-  </Card>
+    </Card>
+  </li>
 );
 
 ExpenseItem.propTypes = {
